Docker Assignment 2/FastAPI-assignment-1-main/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib
bcrypt==4.0.1
python-jose[cryptography]
pydantic
python-multipart